requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.2
openai>=1.3.0
google-generativeai>=0.4.0
pytest>=7.4.0
uvicorn>=0.23.0
fastapi>=0.100.0
rich>=13.7.0
numpy>=1.24.0
pinecone>=5.0.0
langsmith>=0.1.0
duckduckgo-search>=4.0.0

# Optional packages for production
# Uncomment as needed:
# redis>=4.5.0            # For distributed caching
# sqlalchemy>=2.0.0       # For persistent storage
# psycopg2-binary>=2.9.0  # For PostgreSQL backend
